refactor: Convert backup command to command class
